FastApi_ztp_async/requirements.txt
aiofiles==0.6.0
asyncssh==2.5.0
bcrypt==3.2.0
cffi==1.14.5
click==7.1.2
colorama==0.4.4
cryptography==3.4.7
fastapi==0.63.0
future==0.18.2
h11==0.12.0
Jinja2==2.11.3
lxml==4.6.3
MarkupSafe==1.1.1
mypy-extensions==0.4.3
ntc-templates==1.7.0
paramiko==2.7.2
pycparser==2.20
pydantic==1.8.1
PyNaCl==1.4.0
pyserial==3.5
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
scp==0.13.3
scrapli==2021.1.30
scrapli-community==2021.1.30
scrapli-netconf==2021.1.30
six==1.15.0
starlette==0.13.6
tenacity==7.0.0
textfsm==1.1.0
typing-extensions==3.7.4.3
uvicorn==0.13.4
scrapli[asyncssh]
httpx